[ty] Move deferred decorator checks into a submodule
