+++ crates/ty_python_semantic/resources/mdtest/suppressions/no_type_check.md
@@ -53,7 +53,7 @@ def test() -> int:
 
 ## Error in following decorator
 
-Unlike Pyright and mypy, suppress diagnostics in decorator expressions appearing after the
+Unlike Pyright and mypy, we suppress diagnostics in decorator expressions appearing after the
 `no_type_check` decorator. We do this because it more closely matches Python's runtime semantics of
 decorators. For more details, see the discussion on the
 [PR adding `@no_type_check` support](https://github.com/astral-sh/ruff/pull/15122#discussion_r1896869411).
